The database dump is now applied, but poorly.
My main problem with the current strategy is that we attempt to apply
the dump only on first try -- when the database is created. If we fail
some advanced stage after the creation we won't be able to try and
roll out the dump again. Maybe we should create a special file
indicating that the dump has been succesfully applied.