A slightly better initial dump application.
On the very first run of this migration we create the database and
copy the dump file on disk. After the database, all the necessary
setup is done to the database we check if the dump file is on
disk. And if it is we try to apply it and then remove it if we're
successfull.